Server is working except for name
