+++ Lab1/Assets/Scripts/APIWebCall.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: be0478a8cc80067419d8015c67e110dc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
