+++ Lab1/Assets/Scripts/Enemies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b15660385a185a7498208ccc37d7d0b3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
