+++ Lab1/Assets/Scripts/Enemies/EnemyCollision.cs
@@ -4,18 +4,25 @@ using UnityEngine;
 
 public class EnemyCollision : MonoBehaviour
 {
-    int health = 1;
-    private ItemDrop getItem;
-
-    public Animator Anim;
+    private int health;
     private float deathTimer;
+    //private int damageDealt;
+
+    private ItemDrop getItem;
+    public Animator Anim;
+    //public Health playerHealth;
+    
 
     // Start is called before the first frame update
     void Start()
     {
         getItem = GetComponent<ItemDrop>();
         deathTimer = .75f;
+        health = 1;
+        //damageDealt = 1;
 
+       //playerHealth = GameObject.FindWithTag("Player").GetComponent<Health>();
+        
     }
 
     private void Update()
@@ -36,6 +43,11 @@ public class EnemyCollision : MonoBehaviour
                 getItem.DropItem();
                 Debug.Log("Dropped an Item " + getItem);
             }
+        }
+
+        if (collider.CompareTag("Player"))
+        {
+            //playerHealth.takeDamage(damageDealt);
         }
 
     }
