+++ Lab1/Assets/Scripts/GameState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 015733df8d2b9ab49863729addf0bf42
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
