+++ Lab1/Assets/Scripts/Title.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 905ebc109e544b246bc61acb9a197c27
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
