+++ Lab1/Assets/Scripts/Title/Play_Game.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 14e02c20c1ca77b48a5f9eb46d66ca73
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
