+++ Lab1/Assets/Scripts/Title/Player_Name.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9a173d9f19915f478ef3b8d401cc382
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
