+++ Lab1/Assets/Sprites/Title Screen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 71d31d6e3954d084ba3cd29a188981fb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
